2_internal_marks/requirements.txt
Flask
requests
pymongo
mysql-connector-python
flask-cors
